Excluded more compute tests.
