+++ .github/workflows/c2d.yml
@@ -56,6 +56,9 @@ jobs:
       - name: Test with pytest
         run: |
           pytest tests/test_compute.py
+          pytest ocean_provider/utils/test/test_compute.py
+          pytest ocean_provider/utils/test/test_provider_fees.py
+          pytest ocean_provider/validation/test/test_algo_validation.py
       - name:  docker logs
         run: docker logs ocean_aquarius_1 && docker logs ocean_provider_1 && docker logs ocean_provider2_1 && docker logs ocean_ipfs_1
         if: ${{ failure() }}
